Add welcome and error messags
